+++ djexp/export.py
@@ -1,31 +1,21 @@
+import re
+import os
 import sys
 import json
-from re import sub, search
-from os import path, makedirs, listdir, getcwd
+import yaml
 
 from .meta.module import Module
 from .normalizer.normalizers import normalize_root
 
-import yaml
+from .validators import is_valid_module, dir_is_omitted
 
 OUTPUT_FILE = 'django-models'
-
-
-def dir_is_omitted(path_str: str):
-	for item in ['__pycache__', 'venv', 'migrations']:
-		if item in path_str:
-			return True
-	return False
-
-
-def file_id_omitted(file_path: str):
-	return not search('__.+__.py', file_path) and file_path.endswith('.py') and 'setup.py' not in file_path
 
 
 def path_to_module(path_str: str):
 	if path_str.endswith('.py'):
 		path_str = path_str[:len(path_str) - 3]
-	return sub(r'[^\w]+', '.', path_str).lstrip('/.')
+	return re.sub(r'[^\w]+', '.', path_str).lstrip('/.')
 
 
 def save_json(data: dict, target_path: str):
@@ -44,8 +34,8 @@ def save_yaml(data: dict, target_path: str):
 
 def save_dict(data: dict, root_path: str, file_format: str):
 	root_path = normalize_root(root_path)
-	if not path.exists(root_path):
-		makedirs(root_path)
+	if not os.path.exists(root_path):
+		os.makedirs(root_path)
 	target_path = '{}/{}'.format(root_path, OUTPUT_FILE)
 	if file_format is 'json':
 		return save_json(data, target_path)
@@ -55,23 +45,19 @@ def save_dict(data: dict, root_path: str, file_format: str):
 		raise ValueError('invalid serialization file type')
 
 
-def is_valid_module(str_path: str):
-	return path.isfile(str_path) and file_id_omitted(str_path)
-
-
 def get_modules(root: str):
 	modules = []
 	root = normalize_root(root)
-	for f in listdir(root):
-		new_path = path.join(root, f)
+	for f in os.listdir(root):
+		new_path = os.path.join(root, f)
 		if is_valid_module(new_path):
 			modules.append((path_to_module(new_path), new_path))
-		elif path.isdir(new_path) and not dir_is_omitted(new_path):
+		elif os.path.isdir(new_path) and not dir_is_omitted(new_path):
 			modules += get_modules(new_path)
 	return modules
 
 
-def prepare_modules(module_names: [], settings_module: str):
+def prepare_modules(module_names: [], settings_module):
 	modules = []
 	for x in module_names:
 		try:
@@ -90,7 +76,7 @@ def compose_output_data(root_dir: str, modules: []):
 	}, len(final_modules))
 
 
-def export(root_dir: str, settings_module: str, file_format: str = 'json'):
+def export(root_dir: str, settings_module, file_format: str):
 	sys.path.append(root_dir)
 	try:
 		modules = prepare_modules(get_modules(root_dir), settings_module)
@@ -98,12 +84,12 @@ def export(root_dir: str, settings_module: str, file_format: str = 'json'):
 		raise Exception('Error occurred while getting modules\' information: {}'.format(exc))
 	if len(modules) > 0:
 		try:
-			out_data, classes_count = compose_output_data(path.abspath(root_dir), modules)
+			out_data, classes_count = compose_output_data(os.path.abspath(root_dir), modules)
 		except Exception as exc:
 			print(exc)
 			raise Exception('Error occurred while composing output data: {}'.format(exc))
 		try:
-			saved_path = save_dict(out_data, getcwd(), file_format)
+			saved_path = save_dict(out_data, os.getcwd(), file_format)
 		except Exception as exc:
 			raise Exception('Error occurred while json data: {}'.format(exc))
 		print('Exported {} classes, check out \'{}\' file.'.format(classes_count, saved_path))
