+++ djexp/task.py
@@ -1,7 +1,6 @@
 from djexp.export import export
 
 
-def export_django_models(root: str, settings_module: str, file_format: str = None):
-	if file_format:
-		export(root, settings_module, file_format)
-	export(root, settings_module)
+def export_django_models(root: str, settings_module, file_format: str = 'json'):
+	print('Exporting...')
+	export(root, settings_module, file_format)
